requirements.txt
streamlit
python-dotenv
langchain-community
langchain-groq
langchain-google-genai
langchain-ollama